refactor: split into multiple files, cleanup
