+++ CodeEdit.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -77,7 +77,7 @@
       "kind" : "remoteSourceControl",
       "location" : "https://github.com/nathantannar4/Engine",
       "state" : {
-        "revision" : "4a36aef1bec796a0918de81645598668c106f7c4",
+        "revision" : "4ca197be65f7ddeb3387e4e71aed7e508b5b58f5",
         "version" : "0.1.13"
       }
     },
